Split files into model
